FIR - Split of the different exports. etc
